// types.ts
export enum PersonCategory {
  STAFF = 'Staff',
  STUDENT = 'Student',
  PARENT = 'Parent/Guardian',
}

export interface Person {
  id: string;
  category: PersonCategory;
  firstName: string;
  lastName: string;
  image: string; // Base64 encoded image
  role?: string; // For Staff
  class?: string; // For Students
  guardianIds?: string[]; // For Students
  bio: string; // Generated by Gemini
  googleSheetId: string; // Simulated lookup
}